[Refactor] Telemetry 조회 기능 추상화
