+++ src/main/java/com/menu/pubganalyzer/facade/TelemetryFacade.java
@@ -1,0 +1,9 @@
+package com.menu.pubganalyzer.facade;
+
+import com.menu.pubganalyzer.domain.Analyzer;
+import com.menu.pubganalyzer.domain.model.Match;
+import com.menu.pubganalyzer.domain.model.Roster;
+
+public interface TelemetryFacade {
+    Analyzer findLogs(Match match, Roster roster);
+}
